+++ src/main/java/ru/mihalych/randonneuring/telegram/service/TelegramService.java
@@ -6,5 +6,5 @@ public interface TelegramService {
 
      Telegram saveTelegram(Telegram telegram);
 
-     Telegram telegramByTelegramUserChatId(Integer telegramUserId);
+     Telegram telegramByTelegramUserChatId(Long telegramUserId);
 }
